New OpenSTM32 project created
Created an empty C project in OpenST32 in the freertos-stm32 folder that
is this Git clone.

On the 'Target Configuration' page selected 'Series: STM32L0' and
'Board: STM32L0538-DISCO'. On the 'Project Firmware configuration' page
selected 'No firmware' because I'm going to add the firmware manually.

OpenSTM32 spotted it was  Git clone which meant I could right-click the
project in 'Project Explorer' and do 'Team|Add to Index' and then
'Team|Commit...' which brought me to the dialog. 'For kicks' I'm going
to click 'Commit and Push...' rather than 'Commit'.