Enable HAL DMA and SPI modules to suport the Discovery BSP.
The STM32l0538 Discovery BSP makes it easy to use the LEDs on the board.
I want to use the LEDs because I think they'll prove to be a good debug
tool.

I'm not planning on using the DMA or SPI and was hoping to keep to a
minimal set of modules but easy LEDs has overridden that aim.