Added a build varible for the root of FreeRTOS.
The STM32 Cube installation includes a copy of FreeRTOS but, AFAICT, it
is quite old, March 2017, and V9.

I thought it would be more instructive to use the latest version, V10.

I may come to regret this if it turns out there's some special STM32
Cube sauce in the STM32 Cube version.