Created TU for interrupt handlers and define 'SysTick_Handler'.
I copied this file from the STM32 Cube HAL examples and removed all the
empty interrupt handlers which I don't understand.

I really should have created this file before I started working on
FreeRTOS because I think it is important to the HAL. My simple HAL test
worked because it just called a couple of functions and stopped. Without
this definition of 'SysTick_Handler' startup_stm32.s just mapps it to
and inifinte loop so enabling the SysTick wouldn't go well.

My version isn't quite the same as most of the examples. I've added a
call to 'HAL_SYSTICK_IRQHandler' because I saw it in one example. It
just makes it possible to add a /user/ callback, although I'm not sure
what advantage it adds over just adding code to 'SysTick_Handler'.