Added FreeRTOS timers module.
The FreeRTOS timers module isn't necessary for a minimal FreeRTOS
project but it is required for the /normal/ task scheduler (I'm not if
task scheduling can be done in other ways). Using the normal task
scheduler seems like a good starting point for investigating FreeRTOS so
I'm going with timers for now.