Added a second FreeRTOS task to flash the other LED.
This was a bit tricky because I had to increase the heap size and I'm
not sure about the significance of this except that the scheduler
started. I read somewhere in the FreeRTOS instructions that the demo
projects tend to be configured with the minimum heap for that project so
hopefully it just means I copied the config from a single task demo.

As I understand it, I've explicitly created 2 tasks but there is a
third, the idle task, created by the FreeRTOS scheduler.