Use template 'system_stm32l0xx.c' rather than my own /minimal/ one.
I was trying to be clever and minimalist by creating a
'system_stm32l0xx.c' and removing everything from it. But it's origin
has just become clear to me and I think we probably need most of it.

I can now see that 'SystemInit' is called from 'startup_stm32.s' so we
probably should have it. I was getting away without having it because
'startup_stm32.s' contains a /weak/ empty definition.

Unlike the HAL driver files I've linked in a similar way this file
contains some board specific values, i.e. oscillator values. Hence why
the examples tend to make a copy of it. I think linking it like this
helps capture the origin.

There's another copy of it in the STM32L053C8-Discovery templates folder
which I considered using instead, but it is the same and I wanted to
capture the relationship to CMSIS.

Using it hasn't actually made any difference to the behaviour of my
test.