Don't call the FreeRTOS /SysTick/ handler before starting the scheduler.
This is important and it appears I was just /getting away with it/
before.

I changed main and strange things started happening. Eventually I
decided it was because main was taking longer and /SysTick/ was firing
before reaching 'vTaskStartScheduler'.

I didn't use CMSIS-RTOS because I was under the impression it was just
an abstraction layer only only contained /thin/ wrappers. I didn't think
I needed another abstraction layer. As I've just discovered
'osSystickHandler' is not the thinnest of wrappers because it calls
'xTaskGetSchedulerState'.