Clear the command output before processing the command.
If the command being processed doesn't fill the buffer some of the
previous commands output remains. Part of the problem is the heavy
handed transmission of the entire buffer rather than looking for a
string terminator.

The static allocation of the buffer is also interesting. I think I have
to use static allocation at present because the task stack is /minimal/.
Investigating stack sizes could be another day's fun.